requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
psycopg2-binary>=2.9.6
redis>=4.6.0
aiohttp>=3.8.5
python-multipart>=0.0.6
pandas>=2.0.3
openpyxl>=3.1.2
PyPDF2>=3.0.1
tabula-py>=2.7.0
email-validator>=2.0.0
python-dateutil>=2.8.2
starlette~=0.46.0